Added sounds back
- Tidied up the quint API

+++ src/main/java/T145/magistics/api/crafting/RecipeRegistry.java
@@ -7,17 +7,17 @@ import net.minecraft.item.ItemStack;
 
 public class RecipeRegistry {
 
-	private static final HashMap<ItemStack, Short> CRUCIBLE_RECIPES = new HashMap<>();
+	private static final HashMap<ItemStack, Integer> CRUCIBLE_RECIPES = new HashMap<>();
 
 	private RecipeRegistry() {}
 
-	public static void addCrucibleRecipe(ItemStack input, short output) {
+	public static void addCrucibleRecipe(ItemStack input, int output) {
 		CRUCIBLE_RECIPES.put(input, output);
 	}
 
-	public static short getCrucibleOutput(ItemStack stack) {
+	public static int getCrucibleOutput(ItemStack stack) {
 		if (!stack.isEmpty()) {
-			for (Entry<ItemStack, Short> entry : CRUCIBLE_RECIPES.entrySet()) {
+			for (Entry<ItemStack, Integer> entry : CRUCIBLE_RECIPES.entrySet()) {
 				if (ItemStack.areItemStacksEqual(stack, entry.getKey())) {
 					return entry.getValue();
 				}
